Username configuration working
When a user enters the site, a PopUp appears requesting for a username
and doesn't hide until it is set. When the user presses the button, a
signal is emitted from the client to the server to create a username
associated with the socket.id. Once the server confirms the operation,
the signal called "new-user-accepted" is sent and the client app allows
the user to start uploading messages.